+++ interface.go
@@ -1,6 +1,7 @@
 package gnfinder
 
 import (
+	"github.com/gnames/gnfinder/config"
 	"github.com/gnames/gnfinder/ent/output"
 	"github.com/gnames/gnlib/ent/gnvers"
 )
@@ -8,9 +9,9 @@ import (
 type GNfinder interface {
 	Find(data []byte) output.Output
 
-	GetConfig() Config
+	GetConfig() config.Config
 
-	ChangeConfig(opts ...Option) GNfinder
+	ChangeConfig(opts ...config.Option) GNfinder
 
 	GetVersion() gnvers.Version
 }